Adds API URL to OpenVPN server configuration
Adds the ApiUrl property to the OpenVpnServer entity and configuration, allowing to specify the base URL for the DataGateCertManager API calls for each server.

This change ensures that API requests are routed to the correct OpenVPN server instance.
It modifies the CertApiClient and OvpnFileApiClient to use IHttpClientFactory to dynamically create HttpClients with the correct base address based on the server's ApiUrl.
Also includes a database migration to add the new column and seed data with default values.