Refactors OVPN file handling for improved API
Moves OVPN file handling logic from client-side to service-side.

This change refactors the OVPN file management process by introducing an `IOvpnFileApiService` to handle the logic previously residing in the `IOvpnFileApiClient`.
It removes redundant mapping configurations and simplifies the controller by directly utilizing the new service.

Additionally, the `DownloadOvpnFileResponse` is updated to carry the byte array of the file content instead of a filestream and unused models are removed.
The shared models library is also updated to version 1.0.14.