Removes direct Telnet connection to OpenVPN
Refactors the application to remove direct Telnet connections to OpenVPN servers and instead uses the OpenVPN microservice for communication.

This change decouples the application from the OpenVPN server's management interface, improving security and maintainability.

Removes ICommandQueue, CommandQueueManager, TelnetClient, WebSocket integration and moves the command execution to OpenVpnMicroserviceClient.
The application now relies on the microservice to execute commands and retrieve data.