Refactors OpenVPN microservice client handling.
Introduces a factory for creating OpenVPN microservice clients, allowing for more flexible and testable client management.

Replaces the singleton OpenVpnMicroserviceClient with a transient one, obtained through the new factory.

This change also modifies the OpenVpnServerService to pass the OpenVpnServer model to the client to enable more detailed logging and error handling.