Updates default GeoLite DB path
Updates the default path for the GeoLite2 database to reflect the new location within the resources directory. This ensures the application can locate the database file correctly without manual configuration.

Also updates the microservice token service to look for the private and public key files in the resources/certs directory, if not specified in the configuration.