Updates dependencies to latest versions
Bumps Microsoft.EntityFrameworkCore and related packages to version 9.0.6.
Updates OpenVPNGateMonitor.SharedModels to version 1.0.30.

This change ensures compatibility with the latest features and security
updates in the dependency libraries. Additionally, this commit addresses an
issue where the SignalR connection was not properly re-established when the
API URL changed, leading to connection errors. It now checks for API URL
changes and recreates the SignalR connection if necessary.