Adds token-based OVPN file download
Implements the ability to generate and associate tokens with OVPN files for secure and controlled downloads.

The changes introduce new DTOs for representing issued OVPN file tokens and an endpoint for adding OVPN files with associated tokens. This facilitates secure access to client OVPN configurations.

Additionally, it updates the shared models package and xunit.runner.visualstudio.