Updates SharedModels dependency
Updates the OpenVPNGateMonitor.SharedModels package to version 1.0.61 in multiple projects. This ensures that all projects are using the latest version of the shared models, resolving potential compatibility issues and incorporating recent updates or bug fixes.

Also, this change prepares status retrieval to allow for future additions to the data stream for the user dashboard.