Updates HubConnectionState to string type
Refactors the `ConnectionStatusDto` to use a string for the `State` property instead of the `HubConnectionState` enum.
This change is needed because the original enum is part of a SignalR client library, causing versioning and dependency issues.
The `HubConnectionState` enum is removed to avoid these problems.
The package version of `OpenVPNGateMonitor.SharedModels` is bumped to reflect these changes.